Refactoring: extract CallBackWithRegistry
